+++ src/test/aggregate.test.js
@@ -13,6 +13,7 @@ const {
   usersSchema,
   studentsSchema,
   teachersSchema,
+  directorSchema,
 } = require('./mock');
 
 describe('# aggregate() Operations', () => {
@@ -23,6 +24,7 @@ describe('# aggregate() Operations', () => {
     Users,
     Students,
     Teachers,
+    newDirector,
     newTeacher,
     newStudent;
 
@@ -54,14 +56,29 @@ describe('# aggregate() Operations', () => {
       await conn.insertOne('test3', newFile2);
 
       // Discriminator
-      Users = new Model(usersSchema, 'Users');
-      Students = Users.discriminator(studentsSchema, 'Students');
-      Teachers = Users.discriminator(teachersSchema, 'Teachers');
+      const Users = new Model(usersSchema, 'Users');
+      const Students = Users.discriminator(studentsSchema, 'Students');
+      const Teachers = Users.discriminator(teachersSchema, 'Teachers');
+      const Directors = Users.discriminator(directorSchema, 'Directors');
 
       newTeacher = await Teachers.new({
         name: 'Teacher',
         subject: 'Math',
         test: 'both',
+        dept: 'Science',
+      });
+
+      newTeacher2 = await Teachers.new({
+        name: 'Teacher2',
+        subject: 'English',
+        test: 'both',
+        dept: 'Language',
+      });
+
+      newDirector = await Directors.new({
+        name: 'Director',
+        salary: 1000,
+        teachers: [newTeacher._data._id, newTeacher2._data._id],
       });
 
       newStudent = await Students.new({
@@ -69,10 +86,31 @@ describe('# aggregate() Operations', () => {
         term: '1º',
         test: 'both',
         teacher: newTeacher._data._id,
+        marks: 10,
+      });
+
+      newStudent2 = await Students.new({
+        name: 'Student2',
+        term: '2º',
+        test: 'both',
+        teacher: newTeacher._data._id,
+        marks: 8,
+      });
+
+      newStudent3 = await Students.new({
+        name: 'Student3',
+        term: '3º',
+        test: 'one',
+        teacher: newTeacher2._data._id,
+        marks: 5,
       });
 
       await conn.insertOne('test3', newTeacher);
+      await conn.insertOne('test3', newTeacher2);
       await conn.insertOne('test3', newStudent);
+      await conn.insertOne('test3', newStudent2);
+      await conn.insertOne('test3', newStudent3);
+      await conn.insertOne('test3', newDirector);
     } catch (error) {
       throw error;
     }
@@ -122,7 +160,7 @@ describe('# aggregate() Operations', () => {
         {
           $lookup: {
             from: 'Files',
-            let: 'test',
+            error: 'test',
             foreignField: '_id',
             as: 'files',
           },
@@ -139,7 +177,7 @@ describe('# aggregate() Operations', () => {
     } catch (error) {
       expect(error).to.have.property(
         'message',
-        'The paramater let in $lookup aggreagate operator is not allowed',
+        'The paramater error in $lookup aggreagate operator is not allowed',
       );
     }
   });
@@ -193,9 +231,11 @@ describe('# aggregate() Operations', () => {
 
       const aggregate = await conn.aggregate('test3', 'Students', pipeline, {});
 
-      expect(aggregate).to.have.lengthOf(1);
+      expect(aggregate).to.have.lengthOf(2);
       expect(aggregate[0]).to.have.property('name', 'Student');
       expect(aggregate[0]).to.have.property('test', 'both');
+      expect(aggregate[1]).to.have.property('name', 'Student2');
+      expect(aggregate[1]).to.have.property('test', 'both');
     } catch (error) {
       throw error;
     }
@@ -227,11 +267,142 @@ describe('# aggregate() Operations', () => {
 
       const aggregate = await conn.aggregate('test3', 'Teachers', pipeline, {});
 
+      console.log(aggregate);
       expect(aggregate[0]).to.have.property('students');
-      expect(aggregate[0].students).to.have.lengthOf(1);
+      expect(aggregate[0].students).to.have.lengthOf(2);
       expect(aggregate[0].students[0]).to.have.property('name', 'Student');
       expect(aggregate[0].students[0]).to.have.property('test', 'both');
       expect(aggregate[0].students[0]).to.have.property('term', '1º');
+      expect(aggregate[0].students[1]).to.have.property('name', 'Student2');
+      expect(aggregate[0].students[1]).to.have.property('test', 'both');
+      expect(aggregate[0].students[1]).to.have.property('term', '2º');
+    } catch (error) {
+      throw error;
+    }
+  });
+
+  it('performing multiple joins (x2) aggreagation with two discriminator models', async () => {
+    try {
+      const pipeline = [
+        {
+          $lookup: {
+            from: 'Students',
+            let: {
+              teacher_subject: '$subject',
+              teacher_dept: '$dept',
+              teacher_id: '$_id',
+            },
+            pipeline: [
+              {
+                $match: {
+                  $expr: {
+                    $and: [
+                      { $eq: ['$teacher', '$$teacher_id'] },
+                      { $lte: ['$marks', 8] },
+                      { $gte: ['$marks', 5] },
+                    ],
+                  },
+                },
+              },
+            ],
+            as: 'students',
+          },
+        },
+      ];
+
+      const aggregate = await conn.aggregate('test3', 'Teachers', pipeline, {});
+
+      expect(aggregate[0]).to.have.property('students');
+      expect(aggregate[1]).to.have.property('students');
+      expect(aggregate[0].students).to.have.lengthOf(1);
+      expect(aggregate[1].students).to.have.lengthOf(1);
+      expect(aggregate[0].students[0]).to.have.property('name', 'Student2');
+      expect(aggregate[0].students[0]).to.have.property('test', 'both');
+      expect(aggregate[0].students[0]).to.have.property('term', '2º');
+      expect(aggregate[1].students[0]).to.have.property('name', 'Student3');
+      expect(aggregate[1].students[0]).to.have.property('test', 'one');
+      expect(aggregate[1].students[0]).to.have.property('term', '3º');
+    } catch (error) {
+      throw error;
+    }
+  });
+
+  it('performing multiple joins (x3) aggreagation with two discriminator models', async () => {
+    try {
+      const pipeline = [
+        {
+          $lookup: {
+            from: 'Teachers',
+            let: {
+              teachers_id: '$teachers',
+            },
+            pipeline: [
+              {
+                $lookup: {
+                  from: 'Students',
+                  let: {
+                    teacher_id: '$_id',
+                  },
+                  pipeline: [
+                    {
+                      $match: {
+                        $expr: {
+                          $and: [
+                            { $eq: ['$teacher', '$$teacher_id'] },
+                            { $lte: ['$marks', 8] },
+                            { $gte: ['$marks', 5] },
+                          ],
+                        },
+                      },
+                    },
+                  ],
+                  as: 'students',
+                },
+              },
+            ],
+            as: 'teachers',
+          },
+        },
+      ];
+
+      const aggregate = await conn.aggregate(
+        'test3',
+        'Directors',
+        pipeline,
+        {},
+      );
+
+      expect(aggregate[0]).to.have.property('teachers');
+      expect(aggregate[0].teachers[0]).to.have.property('students');
+      expect(aggregate[0].teachers[1]).to.have.property('students');
+      expect(aggregate[0].teachers[0].students).to.have.lengthOf(1);
+      expect(aggregate[0].teachers[1].students).to.have.lengthOf(1);
+
+      expect(aggregate[0].teachers[0].students[0]).to.have.property(
+        'name',
+        'Student2',
+      );
+      expect(aggregate[0].teachers[0].students[0]).to.have.property(
+        'test',
+        'both',
+      );
+      expect(aggregate[0].teachers[0].students[0]).to.have.property(
+        'term',
+        '2º',
+      );
+
+      expect(aggregate[0].teachers[1].students[0]).to.have.property(
+        'name',
+        'Student3',
+      );
+      expect(aggregate[0].teachers[1].students[0]).to.have.property(
+        'test',
+        'one',
+      );
+      expect(aggregate[0].teachers[1].students[0]).to.have.property(
+        'term',
+        '3º',
+      );
     } catch (error) {
       throw error;
     }